twitch_ai_bot/requirements.txt
twitchio
google-generativeai
python-dotenv
